[FEAT] : created MyCarPageView
